+++ matrix.h
@@ -8,7 +8,7 @@ struct Matrix {
 	int m, n;
 	//A list of lists (vectors)
 	//Used to access row vectors individually
-	float **__rows__;
+	double **__rows__;
 
 	//"METHODS"
 	Matrix (*copy)(Matrix self);
@@ -16,9 +16,9 @@ struct Matrix {
 	int *(*dim)(Matrix self);
 	bool (*isNull)(Matrix self);
 	bool (*isOutOfBounds)(Matrix self, int row, int col);
-	float (*getElem)(Matrix self, int row, int col);
-	float *(*getRow)(Matrix self, int row);
-	float *(*getCol)(Matrix self, int col);
+	double (*getElem)(Matrix self, int row, int col);
+	double *(*getRow)(Matrix self, int row);
+	double *(*getCol)(Matrix self, int col);
 	Vector (*getColVector)(Matrix self, int col);
 	Vector (*getRowVector)(Matrix self, int row);
 	bool (*isEqualTo)(Matrix self, Matrix other);
@@ -26,22 +26,22 @@ struct Matrix {
 	//Printers
 	void (*print)(Matrix self);
 	//Setters
-	void (*setElem)(Matrix self, int row, int col, float val);
-	void (*setRow)(Matrix self, int row, float *);
+	void (*setElem)(Matrix self, int row, int col, double val);
+	void (*setRow)(Matrix self, int row, double *);
 	//Operations
-	void (*scale)(Matrix self, float scale);
+	void (*scale)(Matrix self, double scale);
 	void (*transpose)(Matrix *self);
 	//Elementary Operations
 	void (*swapRows)(Matrix self, int row1, int row2);
-	void (*scaleRow)(Matrix self, int row, float scale);
+	void (*scaleRow)(Matrix self, int row, double scale);
 	void (*addRows)(Matrix self, int row1, int row2);
-	void (*addScaledRows)(Matrix self, int row1, float scale1, int row2, float scale2);
+	void (*addScaledRows)(Matrix self, int row1, double scale1, int row2, double scale2);
 };
 
 
 /******************************Constructors******************************/
 Matrix newMatrix(int m, int n);
-Matrix toMatrix(float *, int m, int n);
+Matrix toMatrix(double *, int m, int n);
 Matrix newRandomMatrix(int m, int n);
 
 Matrix __copyMatrix__(Matrix self);
@@ -51,12 +51,12 @@ Matrix __copyMatrix__(Matrix self);
 bool __isNullMatrix__(Matrix self);
 bool __isMatrixOutOfBounds__(Matrix self, int row, int col);
 //Handles accessing the pointer arrays, returns elements
-float __getMatrixElem__(Matrix, int row, int col);
+double __getMatrixElem__(Matrix, int row, int col);
 //These functions will get the row and columns of a matrix as arrays
 //that way they can be operated on more generally instead of being directly converted
 //to vectors
-float *__getRow__(Matrix self, int row);
-float *__getCol__(Matrix self, int col);
+double *__getRow__(Matrix self, int row);
+double *__getCol__(Matrix self, int col);
 //Gets a column of the matrix as a column vector
 Vector __getColVector__(Matrix self, int col);
 //Gets a row of the matrix as a row vector
@@ -72,13 +72,13 @@ void __printMatrix__(Matrix self);
 
 
 /******************************Setters******************************/
-void __setMatrixElem__(Matrix self, int row, int col, float val);
-void __setRow__(Matrix, int row, float *);
+void __setMatrixElem__(Matrix self, int row, int col, double val);
+void __setRow__(Matrix, int row, double *);
 
 
 
 /******************************Operations******************************/
-void __scaleMatrix__(Matrix self, float scale);
+void __scaleMatrix__(Matrix self, double scale);
 void __transposeMatrix__(Matrix *self);
 //Vector multiplication is most general type to decide whether the operation is
 //either an inner product or outer product
@@ -89,11 +89,11 @@ Matrix matrix_mult(Matrix, Matrix);
 
 /******************************Elementary Operations******************************/
 void __swapRows__(Matrix self, int row1, int row2);
-void __scaleRow__(Matrix self, int row, float scale);
+void __scaleRow__(Matrix self, int row, double scale);
 //Will add row2 to row1 and store it in row1
 void __addRows__(Matrix self, int row1, int row2);
 //Will combine scaleRow, and addRows with options of rescaling each row
-void __addScaledRows__(Matrix self, int row1, float scale1, int row2, float scale2);
+void __addScaledRows__(Matrix self, int row1, double scale1, int row2, double scale2);
 
 
 /******************************TESTING FUNCTIONALITY******************************/
